+++ src/modules/controls/orientation/tests/unit/components/poi/PoiOrientation.spec.js
@@ -14,14 +14,19 @@ describe("src/modules/controls/orientation/components/PoiOrientation.vue", () =>
         wrapper,
         returnLegendByStyleIdSpy,
         styleObj,
+        setActiveCategoryStub,
         featureStyleObject;
 
+
     beforeEach(() => {
+        setActiveCategoryStub = sinon.stub();
+
         store = createStore({
             namespaced: true,
             modules: {
                 Controls: {
                     namespaced: true,
+                    state: {},
                     modules: {
                         Orientation: {
                             namespaced: true,
@@ -30,7 +35,7 @@ describe("src/modules/controls/orientation/components/PoiOrientation.vue", () =>
                                 position: sinon.stub()
                             },
                             mutations: {
-                                setActiveCategory: sinon.stub()
+                                setActiveCategory: setActiveCategoryStub
                             }
                         }
                     }
@@ -43,9 +48,9 @@ describe("src/modules/controls/orientation/components/PoiOrientation.vue", () =>
 
         propsData = {
             poiDistances: [
+                500,
                 1000,
-                5000,
-                10000
+                2000
             ],
             getFeaturesInCircle: () => {
                 const feature = new Feature(),
@@ -145,5 +150,11 @@ describe("src/modules/controls/orientation/components/PoiOrientation.vue", () =>
             expect(returnLegendByStyleIdSpy.calledOnce).to.be.true;
         });
 
+    });
+    describe("changedCategory", function () {
+        it("should call setActiveCategory with the category value", function () {
+            wrapper.vm.changedCategory(500);
+            expect(setActiveCategoryStub.calledWith(sinon.match.any, 500)).to.be.true;
+        });
     });
 });